Import of MAGNOLIA/Enbld-0.7026 from CPAN.
gitpan-cpan-distribution: Enbld
gitpan-cpan-version:      0.7026
gitpan-cpan-path:         MAGNOLIA/Enbld-0.7026.tar.gz
gitpan-cpan-author:       MAGNOLIA
gitpan-cpan-maturity:     released

+++ lib/Enbld/Home.pm
@@ -121,10 +121,16 @@ sub _delete_old_build_directory {
 	$limit -= ONE_MONTH;
 
 	foreach my $build_dir ( @list ) {
+        next if ( $build_dir !~ /\d{10}/ ); # for skip .DS_Store
+
 		if ( $limit->epoch > $build_dir ) {
 			remove_tree( File::Spec->catdir( $dirs->{build}, $build_dir ) );
 		}
 	}
+}
+
+sub DESTROY {
+    # do nothing.
 }
 
 1;
